colaborativo mejorado

+++ src/main/java/eci/arsw/eata/model/Location.java
@@ -10,15 +10,15 @@ package eci.arsw.eata.model;
  * @author 2101751
  */
 public class Location {
-    private float latitude;
-    private float longitude;
+    private double latitude;
+    private double  longitude;
 
-    public Location(float latitude, float longitude) {
+    public Location(double  latitude, double  longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
 
-    public float getLatitude() {
+    public double  getLatitude() {
         return latitude;
     }
 
@@ -26,7 +26,7 @@ public class Location {
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+    public double  getLongitude() {
         return longitude;
     }
